Commit late in project
The game is fairly fleshed out at this point. This is fairly late to be
starting the git repo, but better late than never. I just got started
going with creating the game and not bothering to put it on git.

The game has a player controlled boat that can go forward, turn, and fire
cannonballs. Enemies boats have the same capabilities and limitations as
the player boat. The other enemy is an alien. It can abduct crew from the
player's boat. It doesn't hurt the player's health, it just takes away
crew that buff the playerr boat(repair it, increase fire rate,
accelleration).

The map is an open ocean that the player and enemies float around on. The
ocean is a mesh whose height is dictated by a compound sine wave based on
position and the current time. This makes it look like waves on the ocean
(kinda, could be better). All objects have their collider and code at 0 y,
but also have their visual parts float on top of the waves by checking the
value of the sine function at that point.

The game is governed by the Gameplay class. It has a few different
winStates for different parts of the game. The game starts at start with
the player able to move around, but no enemies until they hit start. Then
the elimination mode starts. Enemy pirates are spawned around the player
and the player needs to defeat x of them. Then the next stage begins. A
timer starts counting down and the player needs to survive the spawning
enemies to continue. After the timer runs out, an island is spawned and
the player needs to navigate towards it to finish the game.

Other: The camera follows the player, but at a slightly slower rate. As
such, the player can move away from the camera slightly to be on
diffferent parts of the screen; the player is not always in the very
center of the screen. If the player gets too far away, the camera moves
faster to catch up so the player doesn't get off screen.
When cannonballs hit something to do damage, they pop up damage text that
fades over a short duration. The text is the damage and sometimes a short
message afterwards as a bit of flavour(at least, hopefully it's
interpreted that way).