Aliens now spawn
Added to the spawn code to spawn an alien 20% of the time after the first
survival level. Aliens find the player if they are spawned without getting
told about the player. Aliens don't steal crew if no crew exists (they do
just latch on to the player and follow them forever tho...). Surviving the
night now changes the intensity of the light in the scene so it looks like
nighttime. Added comments and removed redundant code for the island.