Waypoint Arrow
Added a waypoint that can point towards targets of interest such as the
last remaining enemies in survival or where to go in aToB. Added a
function to the camera controller to check if an object is on screen. It
has some wiggle room as it uses the same positions as the speed up code
when the player gets close to the edge. I don't want the waypoint to
disappear as soon as the object is on screen, I'd rather wait until it is
fully on screen to make sure the player sees it. It doesn't work great
when the object and player are both on the bottom of the screen  (the
arrow points right through the object).