Cannonballs disappear, Waypoint arrows spawned
Changed the isOffScreen() method in CameraController slightly. It used to
use the same values that the camera uses to keep the character on screen.
This didn't work great for the arrows in the vertical direction so I added
different values for this.

Made the cannonball float a bit above the waves instead of riding the
waves. Cannonball visual has an arc to it. It appears to dissappear below
the waves as it reaches its max distance. Collider turns off slightly
sooner so cannonballs aren't just floating in the water like mines.

Spawn waypoints to find point of interest. Waypoints spawn in the
elimination game state. Once all enemies are spawned, waypoints point
towards the enemies that are left. A waypoint is spawned when the island
is spawned in aToB stage.
These waypoints orbit the player and point towards the point of interest.
The waypoints hide themselves when the object is on screen; they only show
up to point to something off screen that the player can't see.
By "off screen", I mean the method in CameraController, which is a sort of
approximation of off screen with some liberties taken to look nice. It's
not trying to be accurate.