Flat shading
Added flat shading to the ocean mesh. Turned out to be a dead end. It
didn't look right with the mesh moving along with the camera. Made it look
like the player was stuck in the same box the whole time and that they
weren't really moving. Shading looked cool if the mesh didn't move at the
same time. Would work if I added new meshes as I went.