Abduction messages
The crew left on the boat say somehing when a crew member gets abducted.
Night now goes totally dark in the middle to mask the switching shadows.
Island should now destroy itself when the player dies and not destroy its
gameplay object. Moved the killing of all enemies to its own method so it
caan be called when the player wins.