New Boats
Added new boat models. Added first model without the sail. Made the sail
and tried to import it by just hiding the boat in blender. Didn't work
like that, got the boat again.

Smoothed out sinking and reviving a little. Boat now does the reverse
sinking as the respawn. Modify sinkTime and sinkRate to abuse the same
method to rise and sink. This makes it hard to follow. It could have been
done better.

Added some sounds to when the cannonball does damage. Sounds are supposed
to sound like crashing wood, but they may just sound like generic crashes.
Pick a random of 3 sounds to play.

Moved the deleting of the island out of the killAllEnemies() method. This
way, The final level can call kill all enemies without killing the island.
Island tip message is now said closer to the island. Used to be sais right
as the island spawned. Added a shout message to the end level that gets
called 1 second after the end stage is reached.
Island now destorys the arrow that points to it when it gets destroyed.