Aliens and palm trees
Added a spaceship model. It's simple, but I think it works. Made 1 palm
tree model. I just rotate it and make them diffferent heights to make them
look a little different on the island.

Hopefully fixed a bug where the alien abducts as it dies and the crew is
lost.

Renamed some materials so that they all get imported properly. If they are
all named the same, the new ones don't get imported.