Alien death animation
When the alien dies, it sinks below the water. It also rotates a bit to
look like it lost control. It keeps moving forward a bit to look like it
still has momentum.

When the alien dies, it set alive to false. When alive is false, it can't
attach to the player or set abducting to true on its own. This should stop
the alien from abducting when it's dead.